feat: ApplyStatusCard 컴포넌트를 apply 폴더로 이동 및 수정
